test(tmux): fix CI-fragile sleep, add non-JSON filter test, extract helpers, document dual-gate
- Replace real sleep(80) in out-of-order delivery test with vi.waitFor()
 to avoid flakiness on slow CI (batch-8 issue 35)
- Add test verifying non-JSON files (.log, .txt) ignored by message watcher
 — readFile never called (batch-8 issue 36)
- Extract realExec/isTmuxAvailable into tests/integration/tmux/test-helpers.ts
 using robust version with probe session check (batch-8 issue 37)
- Add test:tmux and test:tmux:integration to npm test warning message
 (batch-8 issue 38)
- Document dual-gate session cap with DESIGN DECISION comments in
 TmuxConnector (in-memory gate) and TmuxSessionManager (tmux-level gate)
 (batch-8 issue 8)
