+++ package.json
@@ -16,7 +16,7 @@
   },
   "scripts": {
     "pretest": "rm -rf test-db test-logs",
-    "test": "echo '\n⚠️  WARNING: Running full test suite crashes Claude Code instances!\n\n✅ Safe commands (run these from Claude Code):\n   npm run test:core\n   npm run test:handlers\n   npm run test:repositories\n   npm run test:adapters\n   npm run test:implementations\n   npm run test:services\n   npm run test:cli\n   npm run test:dashboard\n   npm run test:integration\n\n❌ Full suite: Use npm run test:all (only in local terminal/CI)\n' && exit 1",
+    "test": "echo '\n⚠️  WARNING: Running full test suite crashes Claude Code instances!\n\n✅ Safe commands (run these from Claude Code):\n   npm run test:core\n   npm run test:handlers\n   npm run test:repositories\n   npm run test:adapters\n   npm run test:implementations\n   npm run test:services\n   npm run test:cli\n   npm run test:dashboard\n   npm run test:tmux\n   npm run test:tmux:integration\n   npm run test:integration\n\n❌ Full suite: Use npm run test:all (only in local terminal/CI)\n' && exit 1",
     "test:all": "npm run test:core && npm run test:handlers && npm run test:services && npm run test:repositories && npm run test:adapters && npm run test:implementations && npm run test:cli && npm run test:dashboard && npm run test:scheduling && npm run test:checkpoints && npm run test:error-scenarios && npm run test:orchestration && npm run test:translation && npm run test:integration && npm run test:tmux && npm run test:tmux:integration",
     "test:translation": "NODE_OPTIONS='--max-old-space-size=2048' vitest run tests/unit/translation --no-file-parallelism",
     "test:services": "NODE_OPTIONS='--max-old-space-size=2048' vitest run tests/unit/services/judge-exit-condition-evaluator.test.ts tests/unit/services/task-manager.test.ts tests/unit/services/recovery-manager.test.ts tests/unit/services/process-connector.test.ts tests/unit/services/handler-setup.test.ts tests/unit/services/loop-manager.test.ts tests/unit/services/eval-batch3.test.ts tests/unit/services/eval-batch3-mcp.test.ts tests/unit/services/eval-domain-batch2.test.ts tests/unit/services/schedule-executor-autostart.test.ts tests/unit/services/schedule-executor-pure-fns.test.ts tests/unit/services/eval-task-waiter.test.ts --no-file-parallelism",
